requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aioredis==2.0.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
asyncio==3.4.3
attrs==25.2.0
beautifulsoup4==4.13.3
bitarray==3.1.1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
ckzg==2.0.1
click==8.1.8
cryptography==44.0.2
cytoolz==1.0.1
dnspython==2.7.0
ecdsa==0.19.0
eth-account==0.13.5
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.2.0
eth-typing==5.2.0
eth-utils==5.2.0
eth_abi==5.2.0
fastapi==0.115.11
frozenlist==1.5.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.163.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==1.5.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.1
grpcio==1.71.0
grpcio-status==1.71.0
grpcio-tools==1.71.0
h11==0.14.0
h2==4.2.0
hexbytes==1.3.0
hpack==4.1.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.0.0
loguru==0.7.3
motor==3.7.0
multidict==6.1.0
numpy==2.2.3
packaging==24.2
pandas==2.2.3
parsimonious==0.10.0
pluggy==1.5.0
portalocker==2.10.1
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.3
pyasn1==0.4.8
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pymongo==4.11.2
pyparsing==3.2.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.1
pyunormalize==16.0.0
qdrant-client==1.13.3
redis==5.2.1
regex==2024.11.6
requests==2.32.3
rlp==4.1.0
rsa==4.9
setuptools==76.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.46.1
toolz==1.0.0
tqdm==4.67.1
types-requests==2.32.0.20250306
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
uv==0.6.6
uvicorn==0.34.0
web3==7.8.0
websockets==13.1
yarl==1.18.3
